chore: updae babel and cloudflare libraries
i tried to add this simple test
```
describe('Addition', () => {
  it('knows that 2 and 2 make 4', () => {
    expect(2 + 2).toBe(4);
  });
});
```

but was getting deprecation warnings, so i upgraded babel, i was
also getting a deprecation warning for cloudflare libraries, but alas
one is still left.

i will say babel/webpack is not my strong point, this is one area
where i have slight javascript fatigue, and every company does it
very differently.

Some people may wonder why in a code test, i upgraded everything,
but this is one of the first things i woudl do when i go into any
organisation, upgrading dependencies, tends to show the neglected
parts of the codebase, which haven't been touched in a while.

As a new developer, it tends to show me where all the ghosts in a
cupboard are in a codebase, and also immediately gives an uplift
in developer productivity, as they can use all the current
documentations, and methods.