feat: udpate visual board with bomb generator
i did consider splitting it into rows and columsn, but it is not
strictly needed right now, and doesn't make a difference.

i made an empty Square tpye for the moment, the next step is to fill
the non empty squares with numbers.

in the test, for the moment, i jsut spread the static props, as at
the moment the logic is simple enough, you would split the static props
test out probably but it's enough coverage for now