feat: add bombs nearby squares
a relatively simple implementation, there may be
faster implementations, but i think this is about
as readable as i can get it.

again the sheer nature of the radnomness, makes
this somewhat hard to test, but i am happy with
tha amount of coverage.

given proper input validation, i would be happy
with this level of coverage for mvp