feat: add right click handler + test interactions
The implementation now uses two different reducer(esque) patterns
to deal with user interactions. Due to games inherently having a
lot of shared state, sometimes you have to interact with most of
the state at once. I have dealt with this conundrum before, and
it's a tricky one.

This is why i resort to using cloneDeep, and accepting some mutating
of state in the reducers. In production, you may wish to use something
like immutable.js or a similar pattern, but this is enough for a mvp.
it's mainly about the readability for me in a code test,
this is one of the  parts i was not entirely happy with.

The longest part of this individual feature was the testing, but I am
happy with the level of coverage that it now has.