chore: remove cf-style-nextjs and @cloudflare/style-container
maybe a controversial choice.

@cloudflare/style-container comes in at 195 kB unpacked
cf-style-nextjs comes in at 289KB

emotion and emotion styled come in at 30kb, for this it seems ok

the main reason for me was that, deprecation warnings where showing
and i didn't want deprecations in the console or my tests for a code
test, you can call this, code test driven development.