chore: create infor bar container
there was a mix of logic and presentation, so seperate the two.
i was having trouble tesitng the timers, and didnt want to go
down a rabbit hole for a code test, so left it.

I also didnt seperate out the container logic for the index page
there is a fair amount of testing on that already, and it's simple
enough, just acknowledging that i did think about it, but you have
to stop polishing somewhere.