chore: fix non rectangular boards
rectangular boards wehre rendering incorrectly, this was
due to the parameters being mixed for width and height.

this would of been picked up earlier, if the html closely
represented the data structure (i.e. rows should be nested in
an `<ol>`)

put a test in to make sure this does nto happen again